config: defer opening controls that don't exist at boot
TinyHAL attempts to open all controls during boot to reduce
latency when they are used later. Some controls might not exist
during boot (for example, controls in DSP firmwares loaded
later; accessories that might not be plugged in during boot.)
To handle this a missing control is no longer a fatal error and
an attempt will be made to open it when it is actually used.

Change-Id: I9d1ce1aaf015079454bb6bc84e2b50faf670ddfc
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>