audio: remove the voice recognition trigger handling
This change removes the code that spun off a thread to
call poll() on a compressed stream to wait for a trigger
event from voice recognition. It is much cleaner and easier
to integrate voice control using a totally separate event
from the kernel to indicate the trigger and use the compressed
stream only for transferring the audio. Removing the trigger
thread simplifies the compressed input stream handling.

Change-Id: Ifa3e1da820d691eb5971cd4dd0a55a811e206dba
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>