audio: add voice trigger streams
This adds support for a always-on voice trigger stream.
A voice trigger stream is a hardware pseudo-stream that
sets up some piece of hardware to generate a trigger event
when voice is detected. Defining this as a pseudo-stream
(a stream of type "hw") allow the controls needed to enable
and disable the trigger hardware to be defined in the
TinyHAL configuration file.

Change-Id: I7905439f66d141c2c02ef1fdf7b6d012aa1afa7c
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>