audio: defer opening input until we know input_source
When AudioFlinger opens an input stream the API doesn't
have a way to send the input_source setting - that is
sent later in a set_parameters() call. We must not
open any stream as a default because that stream's config
could break the real stream that will eventually be chosen
by the input_source (particularly if this input was opened
to read a triggered always-on voice recognition stream.)

Instead, the opening of the stream is deferred until the
input_source has been set.

Change-Id: I89c2238c1750fd0b6d8ab869e0ea3876272d61b7
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>