audio: Don't use default sample rates
Having default values for sample rates isn't valid, because
the config file should define the hardware rate and Android
should open a stream with a valid rate before trying to pass
any audio through it. Also, the implementation used for
default rates caused problems when Android is probing a stream
to see what rates it supports because rate that didn't match
the hardware rate was rejected.

Instead, the default rates have been removed and we always
store the actual rate that AudioFlinger passed. But now the
functions to query rate use the hardware rate if AudioFlinger
passed a rate of 0, to support the probing. A similar change is
made for channel mask.

Change-Id: I6adf2d25f02c66be22c4127302451f3de28d68f5
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>