config: store controls by id instead of address
Store the control id and call mixer_get_ctl() when needed
instead of storing the struct mixer_ctl pointer. A call to
mixer_update_ctls() can realloc the list of mixer_ctl structs
which invalidates all our pionters. The mixer_get_ctl() lookup
is a trivial array index so the overhead compared to storing
the pointer is negligible.

Change-Id: I729bd53b295c6ec4bba7a0ab00c44e5c3f8b0266
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>