config: Store controls by pointer with older tinyalsa libraries
The old tinyalsa library shipped with Android Oreo doesn't have
mixer_ctl_get_id() which means we have to store controls by
pointer instead of by id. Add a build define TINYALSA_NO_CTL_GET_ID
to indicate that we cannot use mixer_ctl_get_id().

A consequence of having to store by pointer is that we cannot
update the control list at runtime because that would cause
the pointers to change as the list is realloced in libtinyalsa.

Change-Id: Iab953b2f826215a3965128f94d82e8d9fbca4f90
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>