config: Fix memory leak in parse_stream_start
If parse_stream_start returned an error the memory allocated by the
strdup() would be leaked. Most of the function doesn't need a duplicated
copy of the string, so simplify the fix by only doing the strdup() at the
end of the function when saving a copy of the string.

Change-Id: I765f5e495f99d67b0c91812a509e61f8cfe71ca5
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>