config: Check that we have values in make_byte_array
If strtok() doesn't find any tokens in the value string we should
fail this as an error instead of silently creating a zero-length
array. Also this lack of checking for this case can cause static
analyzers to flag it as a use of zero-length memory since they don't
understand that if the first strtok() returned NULL on the counting
loop it will do the same on the later loop, so the copy loop would
not in fact execute.

Change-Id: If95c2847194554a2c79c19ee9efd39847f2d1630
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>