audio: add audio format return from {in,out}_get_parameters() functions
The Google VTS tests for Android check support for input and output audio
formats using the {in,out}_get_parameters() functions and expect the
functions to write audio formats to the input string (const char *).

Implement the functions {in,out}_get_parameters(), add conditions to
return the input and output audio formats.
Add the get_audio_format() function, which converts the
audio format numeric value to the corresponding format name.

Change-Id: I2c779509b4c37ce5566396a01333102cecd991cc
Signed-off-by: Maksym Prymierov <maksym.prymierov@globallogic.com>