configmgr: Fix include path handling in Linux Makefile
An incorrect path to tinyalsa include directory was hardcoded into the
makefile. Remove this since we expcet that it would normally be on the
CPATH/C_INCLUDE_PATH environment variable that is automatically used by
the compiler.

Also rework the way the list of additional includes is generated so that
extra directories can be passed into the makefile by assigning a makefile
variable EXTRA_C_INCLUDE_PATHS containing a whitespace-separated list.

Change-Id: I4696e667d4c568d299e5cb31e512a9fedd186ce6
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>