configmgr: Improve handling of libraries in Linux Makefile
Use a variable to hold the list of libraries to link against
instead of hardcoding them in the compile recipe.

This list is converted to a set of compiler flags in the same
way as the includes, using $(foreach,...). The variable holding
the resulting flags is named LIB for stylistic consistency with
the existing SRC/OBJ/DEP variables. To accommodate this the
variable holding the target binary is renamed to TRG.

Change-Id: I2d50687481b8b4e6275e110c1f21b7e683b75f85
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>