configmgr: Default to suppressing ALOGV output in Linux builds
The log macros used in a Linux build were resulting in generating
verbose (ALOGV) output by default. This was because the definitions
were derived from Android which set the default for LOG_NDEBUG
based on another preprocessor define NDEBUG. This doesn't exist in
the Linux build and the result was to default LOG_NDEBUG to 0 to
disable supression (that is, enable verbose).

If LOG_NDEBUG has not been defined by the including file it should
default to 1 to suppress verbose messages.

Change-Id: Iacb46737474acaba5fea650b329fd31c1176a335
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>