configmgr: test: Add tests for <ctl> writes to byte controls
Various test cases that <ctl> elements write values to byte controls
correctly.

Each <ctl> is invoked by opening a named stream that then invokes a
global enable path unique to that stream. The mechanism that results
in a stream invoking a path is assumed to work and be tested in some
other test.

To test a reasonable number of combinations means generating a .xml
that contains many cases and large chunks of data. This would be
inconvenient to maintain as a source file so is generated by the test,
together with a corresponding controls .csv and test files of random
data to be loaded into controls.

Change-Id: I2153f143f3ee8535c4476e368bd6e4ed4056083f
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>