configmgr: test: Add tests for <ctl> writes to int controls
Various test cases that <ctl> elements write values to int controls
correctly.

Each <ctl> is invoked by opening a named stream that then invokes a
global enable path unique to that stream. The mechanism that results
in a stream invoking a path is assumed to work and be tested in some
other test.

To test a reasonable number of combinations means generating a .xml
that contains many cases and large chunks of data. This would be
inconvenient to maintain as a source file so is generated by the test,
together with a corresponding controls .csv.

Change-Id: I6e7d64e6b7f9298d0ab2c53fbd05b6e18770a6b3
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>