configmgr: test: Add tests for <ctl> writes to enum controls
Test cases that <ctl> elements write values to enum controls correctly.

Each <ctl> is invoked by opening a named stream that then invokes a
global enable path unique to that stream. The mechanism that results
in a stream invoking a path is assumed to work and be tested in some
other test.

Change-Id: Ie54b6dd2d7cba0e827b48ba0dc048f2b3b43f7d6
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>