configmgr: test: Add tests for <ctl> writes to bool controls
Test cases that <ctl> elements write values to bool controls correctly.

Each <ctl> is invoked by opening a named stream that then invokes a
global enable path unique to that stream. The mechanism that results
in a stream invoking a path is assumed to work and be tested in some
other test.

Change-Id: I0fde0eec331fc19115b8c0e443c35185e32e7dc5
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>