configmgr: test: Add tests for <ctl> elements in a <path>
Test cases that all the <ctl> elements in a <path> are executed and that
execution order is correct.

Each <ctl> is invoked by opening a named stream that then invokes a
global enable path unique to that stream. The mechanism that results in
a stream invoking a path is assumed to work and be tested in some other
test. The correct execution of <ctl> elements is also assumed to work
and be proved elsewhere.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>

Change-Id: I499caad956b664a059543d38410f6f8ea32199b3