configmgr: test: Add test of <device> set reporting and name mapping
Test that get_supported_output_devices() and get_supported_input_devices()
correctly report the set of bits representing the declared <device>
elements. This also tests that the mapping of device names to bitfields is
correct.

This doesn't test every possible combination of devices - that would create
a very large number of tests. Instead it is parameterized so that each pass
through the test cases adds one more output and input device each time. This
is sufficient to prove that each device is correctly mapped to a device bit.

Change-Id: I5043c02118c25ef1ccf123d1c1a102e41843b1b9
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>