configmgr: Fix apply_route() to obey AUDIO_DEVICE_BIT_IN
The logic of apply_route() is to find which devices changed by performing
logical-ANDs between the passed device bits and the currently set devices
in stream.current_devices.

If an input stream is already connected, the input flag bit will be set in
both current_devices and the desired new set of devices, and so will be
stripped as an unchanged bit. The result is that the set of input devices is
changed into a set of output devices (because Android re-uses the same bits
for both and only the input flag distinguishes them). This will then invoke
the change on an output <device> with the same value in bits 0..30 as the
intended input <device>.

This patch restores the AUDIO_DEVICE_BIT_IN flag to the sets of changed bits.

Change-Id: Idbc3e0458b8403128916b0a68241e3f8cd191572
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>