configmgr: test: Add tests for stream attributes
This tests the attribute values parsed out of a <stream> element and
reported in struct hw_stream. The stream type field is tested using
the helper macros.

This is parameterized on all combinations of a few values for each
attribute, with and without each of the optional attributes. Although
this creates a huge number of tests, each one is trivial and quick to
run so there's not much downside to thorough testing of many
attribute combinations.

Change-Id: I00ea1caf15f81f08d4af4a12ab889f1f0fd664b5
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>