configmgr: test: Add test of get_stream()/get_named_stream() matching
Test that get_stream() finds the correct stream based on the specified
attributes and get_named_stream() finds the stream from the name.

Currently get_stream() only matches based on the type (pcm/compressed)
and direction.

Parameterization is used to run a pass with each combination of matchable
attributes and a <usecase> element within the stream invokes a write to
a control unique to that stream.

Change-Id: I440c4156577444bf7debddb903e67d7f62ed0c46
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>