configmgr: Fix get_stream() to only return unnamed streams
The intention of named streams was to be able to fetch a stream for a
specific purpose, where the stream parameters (type, rate, etc) may be
unknown or not unique, but the purpose is unique.

Unnamed streams are intended for normal Android playback/capture streams
where a required {type, direction, rate etc.} is requested and a suitable
stream is found by matching those parameters. In those cases any stream
matching the parameters is considered suitable. It is entirely possible
for a named stream to exist with parameters that would match, but it
should not be considered because it is by definition reserved for
specific purposes.

When named streams were added, the get_stream() logic was not updated to
exclude them. The result was that get_stream() could match a named
stream which was not the intended stream and was probably not appropriate,
and that stream would then not be available to get_named_stream() when
requested for its intended purpose.

This patch fixes get_stream() to only return unnamed streams. To improve
effciency the stream list has been split into two lists, for anonymous
(unnamed) and named.

Change-Id: Id5a70b01478e381158c323a51cc7a130b74662fe
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>