configmgr: Default struct hw_stream.card_number to mixer card number
Defaulting card_number to 0 is ambiguous since it is impossible to tell
whether that means "undefined" or was actually set to zero. If the stream
didn't have a card attribute to override the default, then the card is
the main mixer card number (as specified in the <mixer> element), so
card_number can be defaulted to the actual mixer card number.

Change-Id: I13850f749853863b0f58a0d7fa8551aa212b5212
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>