configmgr: test: Add tests for <codec_probe>
This tests <codec_probe> handling by generating a 3-level set of XML
files and 2 levels of codec_probe and then parameterizing on various
possible ways the codec_probe and probe file could be defined.

The XML path layering is:

A root XML containing a <codec_probe> with several cases, each pointing to
a corresponding  "level-1" XML.

Each level-1 XML has a <codec_probe> with several cases, each pointing to
a corresponding level-2 XML. This means that there are a set of level-2
XMLs for EACH level-1 XML.

To test which XML was finally parsed, all XMLs define a global stream with a
<set> constant. The value of the constant is different for every XML.

Parameterization covers the following tests:
- probe file path absolute or relative
- <case> file path absolute or relative
- <case> target file is in current directory or a subdirectory
- probe file value matches one of the cases
- probe file value does not match any case
- probe file is empty
- probe file value does or doesn't end with newline

All combinations of these are tested. Although it would be possible to assume
that in some cases "testing X implies that Y probably also works", it really
doesn't save much time over testing all combinations.

Change-Id: I7622faf0d2e5ff66ca0d4b2560b90da80813df1f
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>