configmgr: Fix do_parse() to stop properly after a codec_probe
A codec_probe redirection halts parsing of the current XML by calling
XML_StopParser(). But do_parse() was not checking for this. As a result,
a codec_probe only worked if it ended in the last 256-byte chunk of the
XML so that eof=1 when XML_Parse() was called and the while loop in
do_parse() exited when XML_Parse() returned. If it wasn't the last
256-byte chunk of the XML, the while loop would go around and read the
next chunk of file and call XML_Parse() again, which would fail because
parsing has been suspended.

This patch fixes the while loop to check for XML_STATUS_SUSPENDED return
from XML_Parse().

Change-Id: Ie7a0d35013aae8fa64a242c9e98ba00a2828b5b0
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>