configmgr: Fix handling of codec_probe with relative probe file path
If the probe file path specified in a <codec_probe> was not absolute, the
actual location it was loaded from depended on what the owning process
current working directory (getcwd) happened to be when the file was opened.
The owning process could change its cwd and so the location referred to by
a non-absolute path was effectively undefined.

This patch makes non-absolute probe files paths relative to the path of the
XML file containing the <codec_probe> element.

Change-Id: I0e80cc6fdbd019f7bbdfc9b2ace1b3b7fc806c02
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>