configmgr: test: Add test of root XML path handling
Test that the XML file is found correctly when given an absolute
or a relative path to the file.

On Android a relative path is assumed to be relative to the etc
directory (either /vendor/etc or /system/etc, depending on the build
configuration). The main purpose of this test is to prove that this
works. The tests cannot write to etc/ at runtime so the test xml is
pre-installed during build time.

A secondary purpose is to prove that this special Android relative
path handling does not happen on GNU/Linux.

Change-Id: I6577c75df1325b2f10b06c5d1642603b74774c1c
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>