configmgr: Use strtok_r() to prevent memory corruption
strtok() stashes its working pointer somewhere in libc. This is hugely
thread-unsafe. As the configmgr is only a library built into some other
code, it has no way to know whether this stashed pointer is safe. If not,
it can lead to memory corruption if there is another strtok() in action
using the same stash location.

Use strtok_r() instead. It takes a local char** into which it stores its
working pointer.

Change-Id: I4c1846fa0e0c9d54f7675bd59f15cf2380fd445c
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>