configmgr: test: Add testing for memory leaks
Add hooks for memory allocation in test builds of audio_config.c. These
track every allocated address and match it to a free. The primary purpose
is to check for memory leaks, though as an added feature the hooks will
throw an exception (a memory fault on Android) if an address is allocated
without having been freed, or there is a free() for an address that isn't
allocated.

The memory leak checking is added to every existing junit test class after
each call to free_audio_config(). Using the existing tests createa a much
wider range of scenarios than could be reasonably achieved from a dedicated
"memory leak" junit test.

Change-Id: Iba83f8b4789a71eb118daffea64ea0d1924a1826
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>