configmgr: Fix handling of out-of-range volume percentages
If the left volume was out-of-range but the right volume was valid, the
code would lose the error because the right case was executed and
overwrote ret with 0.

Also the illegal values should be trapped immediately, rather than
after possible averaging.

The unnecessary conversion from int to unsigned int made the resulting
error message unhelpful for negative out-of-range values.

Change-Id: I25d2296d72c5cb4436ddc0f0914c211ac3a72fa2
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>